FA-280 Separating the creation page and adding the Curation page to the userflow in Report Management
