+++ frontend/src/components/Sidebar/Sidebar.tsx
@@ -191,13 +191,6 @@ const Sidebar: React.FC<SidebarProps> = ({ isCollapsed, setIsCollapsed }) => {
                     to: "/details-settings",
                     tiers: ["Basic + Financial Assistant", "Custom + Financial Assistant", "Premium + Financial Assistant"],
                     roles: ["admin", "user"]
-                },
-                {
-                    title: "Curation Reports",
-                    icon: <IconClipboardText className={styles.sidebarLinkIcon} />,
-                    to: "/curation-reports",
-                    tiers: ["Basic + Financial Assistant", "Custom + Financial Assistant", "Premium + Financial Assistant"],
-                    roles: ["admin"]
                 }
             ]
         },
