+++ frontend/src/pages/reports/CurationReports.tsx
@@ -1,25 +1,16 @@
 import React, { useEffect, useState } from "react";
 import styles from "./CurationReports.module.css"
-import { Dropdown, Label, ResponsiveMode, Spinner } from "@fluentui/react";
-import { IconFilePlus, IconTrash, IconX } from "@tabler/icons-react";
-import { createReport, deleteReport, getFilteredReports } from "../../api";
+import { Label, Spinner } from "@fluentui/react";
+import { IconArrowBack, IconFilePlus, IconTrash, IconX } from "@tabler/icons-react";
+import { deleteReport, getFilteredReports } from "../../api";
+import { useNavigate } from "react-router-dom";
 
 const CurationReports  = () => {
 
+    const navigate = useNavigate();
     const [dataLoad, setDataLoad] = useState(false)
     const [filteredReports, setFilteredReports] = useState<any[]>([]);
     const [loading, setLoading] = useState(false);
-    const [isCreateModalOpen, setIsCreateModalOpen] = useState(false);
-    const [isConfirm, setIsConfirm] = useState(false);
-    const [inputReportName, setinputReportName] = useState('')
-    const curationReportOptions = [
-        { key: "1", text: "Ecommerce" },
-        { key: "2", text: "Weekly Economic" },
-        { key: "3", text: "Monthly Economic" },
-    ];
-    const [categorySelection, setCategorySelection] = useState('')
-    const [errorMessage, setErrorMessage] = useState<string | null>("");
-    const [isPopupActive, setIsPopupActive] = useState(false)
     const [isDeleteActive, setIsDeleteActive] = useState(false)
     const [deletedReportID, setdeletedReportID] = useState('');
     const [deletedReportName, setdeletedReportName] = useState('');
@@ -50,69 +41,6 @@ const CurationReports  = () => {
         
     }, [dataLoad]);
 
-    const handleCreateButton = () => {
-        setIsCreateModalOpen(!isCreateModalOpen);
-        setCategorySelection('')
-        setinputReportName('')
-        setErrorMessage(null)
-    }
-
-    const handleConfirmButton = () => {
-
-        if(inputReportName == ('')){
-            setErrorMessage('Please type the Name of the Report')
-            return;
-        }
-        if(categorySelection == ('')){
-            setErrorMessage('Please select the Report Category')
-            return;
-        }
-        setIsCreateModalOpen(false);
-        setIsConfirm(!isConfirm);
-    }
-
-    const handleCancelButton = () => {
-        setIsConfirm(false)
-        setIsCreateModalOpen(false)
-        setCategorySelection('')
-        setinputReportName('')
-        setErrorMessage(null)
-    }
-
-    const handleTypeDropdownChange = (event: any, selectedOption: any) => {
-        setCategorySelection(selectedOption.text)
-    }
-
-    const handleInputName = (event: React.ChangeEvent<HTMLInputElement>) => {
-        setinputReportName(event.target.value)
-    }
-    
-    const handleCreateReport = async () => {
-        setIsConfirm(false)
-        setIsCreateModalOpen(false)
-        let timer: NodeJS.Timeout;
-
-        try{
-            await createReport({ 
-                type: "curation",
-                name: inputReportName,
-                category: categorySelection,
-                status: "archived"
-            });
-            setDataLoad(!dataLoad)
-
-            setIsPopupActive(true)
-            timer = setTimeout(() => {
-                setIsPopupActive(false);
-            }, 3000);
-
-        } catch (error){
-            console.error("Error trying to create the report: ", error);
-        }finally{
-            setLoading(false);
-        }
-    }
-
     const handleDeleteButton = (reportID: string, reportName: string) => {
         setIsDeleteActive(!isDeleteActive);
         setdeletedReportID(reportID)
@@ -139,17 +67,22 @@ const CurationReports  = () => {
     }
 
     return (
+    
         <div className={styles.page_container}>
+            <div className={styles.labelContainer}>
+                <button className={styles.button} title="Return to Report Management" aria-label="Return to Report Management" onClick={() => navigate('/view-manage-reports')}>
+                    <IconArrowBack className={styles.iconColor}/>
+                    <Label className={styles.textButton}>Return to Report Management</Label>
+                </button>
+            </div>
             <div id="options-row" className={styles.row}>
                 <h1 className={styles.title}>Curation Reports</h1>
             </div>
             <div className={styles.card}>
                 <div className={styles.labelContainer}>
-                    <Label className={styles.text}>Add a New Curation Report</Label>
-                </div>
-                <div className={styles.buttonContainer}>
-                    <button className={styles.button} title="Add a New Report" aria-label="Add a New Report" onClick={handleCreateButton}>
-                        <IconFilePlus className={styles.iconLarge}/>
+                    <button className={styles.button} title="Add a New Report" aria-label="Add a New Report" onClick={() => navigate('/create-curation-report')}>
+                        <IconFilePlus className={styles.iconColor}/>
+                        <Label className={styles.textButton}>Add a New Curation Report</Label>
                     </button>
                 </div>
                 {loading ? (
@@ -209,60 +142,16 @@ const CurationReports  = () => {
                         </table>
                     )}
             </div>
-            {isCreateModalOpen && (
-                <div className={styles.modal}>
-                    <button className={styles.closeButton} onClick={handleCreateButton}><IconX /></button>
-                    <Label className={styles.text}>Create a Curation Report</Label>
-                    <div>
-                        <form>
-                            <Label>Report Name</Label>
-                            <input type="text" className={styles.input} onChange={handleInputName} value={inputReportName}></input>
-                            <Label>Curation Report Category</Label>
-                            <Dropdown placeholder="Select a Curation Report Category" options={curationReportOptions} onChange={handleTypeDropdownChange} defaultValue={categorySelection} responsiveMode={ResponsiveMode.unknown}/>
-                        </form>
-                    </div>
-                        <div>
-                            {errorMessage !== null && <p className={styles.error}>{errorMessage}</p>}
-                            <div className={styles.buttonContainer}>
-                                <button className={styles.button} title="Confirm" aria-label="Confirm" onClick={handleConfirmButton}>
-                                    Confirm
-                                </button>
-                                <button className={styles.button} title="Cancel" aria-label="Cancel" onClick={handleCancelButton}>
-                                    Cancel
-                                </button>
-                            </div>
-                        </div>
-                </div>
-                
-            )}
-            {isConfirm && (
-                <div className={styles.modal}>
-                    <Label className={styles.text}>Are you sure you want to create the report {inputReportName} ?</Label>
-                    <div className={styles.buttonContainer}>
-                        <button className={styles.button} title="Confirm" aria-label="Confirm" onClick={handleCreateReport}>
-                            Confirm
-                        </button>
-                        <button className={styles.button} title="Cancel" aria-label="Cancel" onClick={handleCancelButton}>
-                            Cancel
-                        </button>
-                    </div>
-                </div>
-            )}
-            {isPopupActive && (
-                <div className={styles.modal}>
-                    <Label className={styles.text}>The report has been added. It needs to be implemented for generation. <br /> 
-                    If not implemented, it will not be generated.</Label>
-                </div>
-            )}
             {isDeleteActive && (
                 <div className={styles.modal}>
+                    <button className={styles.closeButton} onClick={handleCancelDelete}><IconX/></button>
                     <Label className={styles.text}>Are you sure you want to delete {deletedReportName}</Label>
                     <div className={styles.buttonContainer}>
-                        <button className={styles.button} title="Confirm" aria-label="Confirm" onClick={handleConfirmDelete}>
-                                Confirm
-                        </button>
                         <button className={styles.button} title="Cancel" aria-label="Cancel" onClick={handleCancelDelete}>
                             Cancel
+                        </button>
+                        <button className={styles.button} title="Confirm" aria-label="Confirm" onClick={handleConfirmDelete}>
+                                Confirm
                         </button>
                     </div>
                 </div>
