+++ frontend/src/pages/reports/ReportCreation/CurationCreation.module.css
@@ -22,52 +22,9 @@
     margin-bottom: 10px;
 }
 
-.elementrow{
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: center;
-    justify-items: center;
-    margin-bottom: 10px;
-    height: 40px;
-    align-items: center;
-    margin-top: 30px;
-}
-
 .title {
     margin-bottom: 20px;
     margin-left: 20px;
-}
-
-.card {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    background-color: #f5f5f5;
-    border: 1px solid #e3e3e3;
-    border-radius: 6px;
-    box-shadow: 0px 4px 12px 0px rgba(0, 0, 0, 0.07), 0px 2px 4px 0px rgba(0, 0, 0, 0.05);
-    padding: 20px 20px 30px 20px;
-    gap: 20px;
-    margin-top: 15px;
-    flex-grow: 2;
-}
-
-.iconColor{
-    color: #a7c444;
-}
-
-.iconLarge{
-    color: #a7c444;
-    width: 3em;
-    height: 3em;
-}
-
-.container {
-    display: flex;
-    justify-content: center;
-    gap: 300px;
-    width: 100%;
 }
 
 .labelContainer{
@@ -76,19 +33,33 @@
     width: 100%;
 }
 
-.text{
-    text-align: center;
+.card {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    background-color: #f5f5f5;
+    border: 1px solid #e3e3e3;
+    border-radius: 6px;
+    box-shadow: 0px 4px 12px 0px rgba(0, 0, 0, 0.07), 0px 2px 4px 0px rgba(0, 0, 0, 0.05);
+    padding: 20px 20px 30px 20px;
+    gap: 50px;
+    margin-top: 15px;
+    flex-grow: 2;
 }
 
-.textButton{
-    text-align: center;
-    cursor: pointer;
-    margin-left: 10px;
-}
+@media (max-width: 768px) {
+    .title{
+        margin-bottom: 10px;
+        margin-left: 0px;
+    }
 
-.textPanel{
-    text-align: center;
-    margin-bottom: 10px;
+    .center {
+        margin-left: 20px;
+    }
+
+    .info{
+        overflow-x: auto;
+    }
 }
 
 .button{
@@ -103,6 +74,12 @@
     background-color: #ffffff;
 }
 
+.textButton{
+    text-align: center;
+    cursor: pointer;
+    margin-left: 10px;
+}
+
 .button:hover{
     box-shadow: 0 0 8px rgba(139, 195, 74, 0.5);
 }
@@ -111,53 +88,49 @@
     background-color: rgba(139, 195, 74, 0.5);
 }
 
-.buttonPanel{
+.buttonContainer{
+    text-align: left;
+    border-collapse: collapse;
+    width: 100%;
+    justify-content: flex-start;
+    display: flex;
+    gap: 40px;
+}
+
+.iconColor{
+    color: #a7c444;
+}
+
+.input{
     display: flex;
     padding: 5px 10px;
     border: 2px solid #9fc51d;
     border-radius: 6px;
-    height: 40px;
+    height: auto;
+    width: 25em;
     align-items: center;
     justify-content: center;
     background-color: #ffffff;
 }
 
-.buttonPanel:hover{
+.input:focus{
+    border: 2px solid #8eb11a;
+    outline: none;
     box-shadow: 0 0 8px rgba(139, 195, 74, 0.5);
 }
 
-.buttonPanel:active{
-    background-color: rgba(139, 195, 74, 0.5);
-}
-
-.buttonContainer{
-    text-align: left;
-    border-collapse: collapse;
-    width: 100%;
-    justify-content: center;
-    display: flex;
-    gap: 40px;
-}
-
-@media (max-width: 900px){
-    .buttonContainer{
-        gap: 10px;
+@media (max-width: 900px) {
+    .input{
+        width: 16em;
     }
 }
 
-.table {
-    text-align: left;
-    background-color: white;
-    border-collapse: collapse;
-    width: 100%;
+.error {
+    text-align: center;
+    font-weight: 400;
+    font-style: italic;
+    color: #ff3333;
 }
-
-@media (max-width: 900px){
-    .table{
-        font-size: 0.7em;
-    }
-}
-
 
 .modal{
     position: fixed;
@@ -179,13 +152,48 @@
     gap: 30px;
 }
 
-@media (max-width: 900px) {
+.modalPopup{
+    position: fixed;
+    top: 10%;
+    left: 50%;
+    transform: translateX(-50%);
+    width: auto; 
+    height: auto;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center; 
+    z-index: 9999;
+    padding: 100px; 
+    background-color: #cef1a5;
+    border: 1px solid #87f083;
+    border-radius: 6px;
+    box-shadow: 0px 4px 12px rgba(0, 0, 0, 0.07), 0px 2px 4px rgba(0, 0, 0, 0.05);
+    gap: 30px;
+}
+
+@media (max-width: 900px){
     .modal{
         width: 20em;
+        height: 10em;
+        padding: 20px;
     }
-    .info{
-        width: 10em;
+    .modalPopup{
+        width: 20em; 
     }
+}
+
+.buttonModalContainer{
+    text-align: left;
+    border-collapse: collapse;
+    width: 100%;
+    justify-content: center;
+    display: flex;
+    gap: 40px;
+}
+
+.text{
+    text-align: center;
 }
 
 .closeButton {
@@ -205,65 +213,4 @@
 .closeButtonContainer:active{
     background-color: rgba(214, 214, 214, 0.315);
     border-radius: 5px;
-}
-
-@media (max-width: 768px) {
-    .title{
-        margin-bottom: 10px;
-        margin-left: 0px;
-    }
-
-    .center {
-        margin-left: 20px;
-    }
-
-    .info{
-        overflow-x: auto;
-    }
-}
-
-.thead{
-    background-color: #a7c444;
-    color: white;
-}
-
-.tableName{
-    padding: 10px;
-}
-
-.tableText{
-    text-align: justify;
-}
-
-.tableBackground{
-    background-color: white;
-}
-
-.tableBackgroundAlt{
-    background-color: #e9e9e9;
-}
-
-.tableStatusContainer{
-    width: 100%;
-    justify-content: flex-start;
-    justify-items: center;
-    text-align: center;
-    display: flex;
-    text-transform: capitalize;
-}
-
-.tableStatusActive{
-    width: 100px;
-    background-color: #d2ffd2;
-    padding: 5px;
-    color: #a0df2c;
-    border-radius: 15px;
-}
-
-.tableStatusArchived{
-    width: 100px;
-    background-color: #c6f0ee;
-    padding: 5px;
-    color: #2ca6df;
-    border-radius: 15px;
 }
